refactor(structure): reorganize firmware modules by domain
