+++ S-ADAPT/Core/Inc/input/encoder_input.h
