+++ S-ADAPT/Core/Inc/input/input_utils.h
