+++ S-ADAPT/Core/Inc/sensors/ldr.h
