+++ S-ADAPT/Core/Inc/support/debug_print.h
