+++ S-ADAPT/Core/Src/support/debug_print.c
@@ -1,4 +1,4 @@
-#include "debug_print.h"
+#include "support/debug_print.h"
 
 #include <stdarg.h>
 #include <stdio.h>
